aep: Refactor SrgSpec
