+++ emissions/aep/surrogate.go
@@ -21,9 +21,7 @@ package aep
 import (
 	"context"
 	"fmt"
-	"math"
 	"runtime"
-	"strconv"
 	"strings"
 	"sync"
 
@@ -264,171 +262,6 @@ func (g *GriddedSrgData) WriteToShp(s *shp.Encoder) error {
 		}
 	}
 	return nil
-}
-
-// get surrogate shapes and weights. tol is a geometry simplification tolerance.
-func (srg *SrgSpec) getSrgData(gridData *GridDef, inputLoc *Location, tol float64) (*rtree.Rtree, error) {
-	srgShp, err := shp.NewDecoder(srg.WEIGHTSHAPEFILE)
-	if err != nil {
-		return nil, err
-	}
-	defer srgShp.Close()
-
-	srgSR, err := srgShp.SR()
-	if err != nil {
-		return nil, err
-	}
-
-	srgCT, err := srgSR.NewTransform(gridData.SR)
-	if err != nil {
-		return nil, err
-	}
-
-	gridSrgCT, err := gridData.SR.NewTransform(srgSR)
-	if err != nil {
-		return nil, err
-	}
-
-	// Calculate the area of interest for our surrogate data.
-	inputShapeT, err := inputLoc.Reproject(srgSR)
-	if err != nil {
-		return nil, err
-	}
-	inputShapeBounds := inputShapeT.Bounds()
-	srgBounds := inputShapeBounds.Copy()
-	for _, cell := range gridData.Cells {
-		cellT, err := cell.Transform(gridSrgCT)
-		if err != nil {
-			return nil, err
-		}
-		b := cellT.Bounds()
-		if b.Overlaps(inputShapeBounds) {
-			srgBounds.Extend(b)
-		}
-	}
-
-	var fieldNames []string
-	if srg.FilterFunction != nil {
-		fieldNames = append(fieldNames, srg.FilterFunction.Column)
-	}
-	if srg.WeightColumns != nil {
-		fieldNames = append(fieldNames, srg.WeightColumns...)
-	}
-	srgData := rtree.NewTree(25, 50)
-	var recGeom geom.Geom
-	var data map[string]string
-	var keepFeature bool
-	var featureVal string
-	var size float64
-	var more bool
-	for {
-		recGeom, data, more = srgShp.DecodeRowFields(fieldNames...)
-		if !more {
-			break
-		}
-
-		if !recGeom.Bounds().Overlaps(srgBounds) {
-			continue
-		}
-
-		if srg.FilterFunction == nil {
-			keepFeature = true
-		} else {
-			// Determine whether this feature should be kept according to
-			// the filter function.
-			keepFeature = false
-			featureVal = strings.TrimSpace(fmt.Sprintf("%v", data[srg.FilterFunction.Column]))
-			for _, filterVal := range srg.FilterFunction.Values {
-				switch srg.FilterFunction.EqualNotEqual {
-				case "NotEqual":
-					if featureVal != filterVal {
-						keepFeature = true
-					}
-				default:
-					if featureVal == filterVal {
-						keepFeature = true
-					}
-				}
-			}
-		}
-		if keepFeature && recGeom != nil {
-			srgH := new(srgHolder)
-			srgH.Geom, err = recGeom.Transform(srgCT)
-			if err != nil {
-				return srgData, err
-			}
-			if tol > 0 {
-				switch srgH.Geom.(type) {
-				case geom.Simplifier:
-					srgH.Geom = srgH.Geom.(geom.Simplifier).Simplify(tol)
-				}
-			}
-			if len(srg.WeightColumns) != 0 {
-				weightval := 0.
-				for i, name := range srg.WeightColumns {
-					var v float64
-					vstring := data[name]
-					if strings.Contains(vstring, "\x00\x00\x00\x00\x00\x00") || strings.Contains(vstring, "***") || vstring == "" {
-						// null value
-						v = 0.
-					} else {
-						v, err = strconv.ParseFloat(data[name], 64)
-						if err != nil {
-							return srgData, fmt.Errorf("aep.getSrgData: shapefile %s column %s, %v", srg.WEIGHTSHAPEFILE, name, err)
-						}
-						v = math.Max(v, 0) // Get rid of any negative weights.
-					}
-					weightval += v * srg.WeightFactors[i]
-				}
-				switch srgH.Geom.(type) {
-				case geom.Polygonal:
-					size = srgH.Geom.(geom.Polygonal).Area()
-					if size == 0. {
-						if tol > 0 {
-							// We probably simplified the shape down to zero area.
-							continue
-						} else {
-							// TODO: Is it okay for input shapes to have zero area? Probably....
-							continue
-							//err = fmt.Errorf("Area should not equal "+
-							//	"zero in %v", srg.WEIGHTSHAPEFILE)
-							//return srgData, err
-						}
-					} else if size < 0 {
-						panic(fmt.Errorf("negative area: %g, geom:%#v", size, srgH.Geom))
-					}
-					srgH.Weight = weightval / size
-				case geom.Linear:
-					size = srgH.Geom.(geom.Linear).Length()
-					if size == 0. {
-						err = fmt.Errorf("Length should not equal "+
-							"zero in %v", srg.WEIGHTSHAPEFILE)
-						return srgData, err
-					}
-					srgH.Weight = weightval / size
-				case geom.Point:
-					srgH.Weight = weightval
-				default:
-					err = fmt.Errorf("aep: in file %s, unsupported geometry type %#v",
-						srg.WEIGHTSHAPEFILE, srgH.Geom)
-					return srgData, err
-				}
-			} else {
-				srgH.Weight = 1.
-			}
-			if srgH.Weight < 0. || math.IsInf(srgH.Weight, 0) ||
-				math.IsNaN(srgH.Weight) {
-				err = fmt.Errorf("Surrogate weight is %v, which is not acceptable.", srgH.Weight)
-				return srgData, err
-			} else if srgH.Weight != 0. {
-				srgData.Insert(srgH)
-			}
-		}
-	}
-	if srgShp.Error() != nil {
-		return nil, fmt.Errorf("in file %s, %v", srg.WEIGHTSHAPEFILE, srgShp.Error())
-	}
-	return srgData, nil
 }
 
 func genSrgWorker(singleShapeChan, griddedSrgChan chan *GriddedSrgData,
